% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparative_Trials.R
\name{arm_copula_pfs}
\alias{arm_copula_pfs}
\title{Compute copula PFS curve for one arm}
\usage{
arm_copula_pfs(arm, n_times, copula_family, threshold = 1.2)
}
\description{
Compute copula PFS curve for one arm
}
