% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_and_Event_Generation.R
\name{compute_new_lesion_probability}
\alias{compute_new_lesion_probability}
\title{Compute probability of new lesion appearance}
\usage{
compute_new_lesion_probability(coefficients, tumour_sizes)
}
\arguments{
\item{coefficients}{List returned by generate_coefficients()}

\item{tumour_sizes}{Matrix of previous tumour sizes (n × T)}
}
\value{
Matrix of probabilities p_it
}
\description{
Compute probability of new lesion appearance
}
