% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_and_Event_Generation.R
\name{generate_binary_data}
\alias{generate_binary_data}
\title{Generate binary new-lesion indicators from probabilities}
\usage{
generate_binary_data(probabilities)
}
\arguments{
\item{probabilities}{Matrix (n × T) of probabilities p_it}
}
\value{
Matrix (n × T) of 0/1 lesion indicators
}
\description{
Generate binary new-lesion indicators from probabilities
}
