% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{get_true_rates}
\alias{get_true_rates}
\title{Gets "true" survival rates through a Kaplan-Meier estimate on 100,000 patients}
\usage{
get_true_rates(
  n_times,
  n_true_patients = 1e+05,
  mean,
  covariance,
  alpha,
  beta,
  gamma,
  R,
  threshold = 1.2
)
}
\arguments{
\item{n_times}{Integer. Number of follow‑up time points.}

\item{n_true_patients}{Integer. Number of patients to simulate. (default 100,000)}

\item{mean}{Numeric vector. Mean vector for the multivariate normal
distribution of log tumour‑size ratios.}

\item{covariance}{Numeric matrix. Covariance matrix for the
multivariate normal distribution.}

\item{alpha}{Numeric scalar. Baseline intercept parameter for the
logistic new‑lesion model.}

\item{beta}{Numeric scalar. Treatment‑effect parameter for the
logistic new‑lesion model.}

\item{gamma}{Numeric scalar. Effect of previous tumour size in the
logistic new‑lesion model.}

\item{R}{Numeric scalar or vector. Treatment indicator(s) for
patients; defaults to 0 (control arm).}

\item{threshold}{Numeric. Threshold multiplier for defining tumour growth progression (default 1.2).}
}
\description{
Gets "true" survival rates through a Kaplan-Meier estimate on 100,000 patients
}
