% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparative_Trials.R
\name{km_log_rank_test}
\alias{km_log_rank_test}
\title{Perform the log rank test for the control and treatment arms}
\usage{
km_log_rank_test(control_data, treatment_data)
}
\arguments{
\item{control_data}{time and event data for the control arm in the form for KM estimate}

\item{treatment_data}{time and event data for the treatment arm in the form for KM estimate}
}
\value{
p-value for the log rank test
}
\description{
Perform the log rank test for the control and treatment arms
}
