% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparative_Trials.R
\name{logrank_power}
\alias{logrank_power}
\title{Power of log-rank test using the same data-generation framework}
\usage{
logrank_power(
  n_times,
  n_patients,
  n_iterations,
  mean_ctrl,
  mean_trt,
  covariance,
  alpha_coef,
  beta_ctrl = 0,
  beta_trt,
  gamma,
  alpha_level = 0.05,
  threshold = 1.2,
  seed = NULL
)
}
\arguments{
\item{n_times, n_patients, n_iterations, mean_ctrl, mean_trt, covariance, }{alpha_coef,beta_ctrl,beta_trt,gamma,threshold,seed  same as frank_copula_power}
}
\value{
list: power, p_values
}
\description{
Power of log-rank test using the same data-generation framework
}
