% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumourSim-package.R
\docType{package}
\name{tumourSim-package}
\alias{tumourSim}
\alias{tumourSim-package}
\title{tumourSim: Simulation Tools for Tumour Progression Models}
\description{
Simulation framework for log tumour size ratios and lesion probabilities. The true survival rates, Kaplan-Meier estimated survival rates and Copula estimated survival rates and metrics are calculated.
}
\author{
\strong{Maintainer}: Calum Regan \email{c_regan1888@outlook.com}

}
